fix float to double in npb_similar_player, network_intrusion, rent

+++ npb_similar_player/cpp/update.cpp
@@ -37,24 +37,24 @@ int main(int argc, char* argv[]){
         >> obp >> ops >> rc27 >> xr27;
 
     d.add_string("チーム", team);
-    d.add_number("打率", lexical_cast<float>(bave));
-    d.add_number("試合数", lexical_cast<float>(games));
-    d.add_number("打席", lexical_cast<float>(pa));
-    d.add_number("打数", lexical_cast<float>(atbat));
-    d.add_number("安打", lexical_cast<float>(hit));
-    d.add_number("本塁打", lexical_cast<float>(homerun));
-    d.add_number("打点", lexical_cast<float>(runsbat));
-    d.add_number("盗塁", lexical_cast<float>(stolen));
-    d.add_number("四球", lexical_cast<float>(bob));
-    d.add_number("死球", lexical_cast<float>(hbp));
-    d.add_number("三振", lexical_cast<float>(strikeout));
-    d.add_number("犠打", lexical_cast<float>(sacrifice));
-    d.add_number("併殺打", lexical_cast<float>(dp));
-    d.add_number("長打率", lexical_cast<float>(slg));
-    d.add_number("出塁率", lexical_cast<float>(obp));
-    d.add_number("OPS", lexical_cast<float>(ops));
-    d.add_number("RC27", lexical_cast<float>(rc27));
-    d.add_number("XR27", lexical_cast<float>(xr27));
+    d.add_number("打率", lexical_cast<double>(bave));
+    d.add_number("試合数", lexical_cast<double>(games));
+    d.add_number("打席", lexical_cast<double>(pa));
+    d.add_number("打数", lexical_cast<double>(atbat));
+    d.add_number("安打", lexical_cast<double>(hit));
+    d.add_number("本塁打", lexical_cast<double>(homerun));
+    d.add_number("打点", lexical_cast<double>(runsbat));
+    d.add_number("盗塁", lexical_cast<double>(stolen));
+    d.add_number("四球", lexical_cast<double>(bob));
+    d.add_number("死球", lexical_cast<double>(hbp));
+    d.add_number("三振", lexical_cast<double>(strikeout));
+    d.add_number("犠打", lexical_cast<double>(sacrifice));
+    d.add_number("併殺打", lexical_cast<double>(dp));
+    d.add_number("長打率", lexical_cast<double>(slg));
+    d.add_number("出塁率", lexical_cast<double>(obp));
+    d.add_number("OPS", lexical_cast<double>(ops));
+    d.add_number("RC27", lexical_cast<double>(rc27));
+    d.add_number("XR27", lexical_cast<double>(xr27));
 
     r.update_row(pname, d);
   }
